Email verification sent without entering the app

+++ src/components/app/app.js
@@ -24,6 +24,8 @@ const App = () => {
 
     const [signInType, setSignInType] = useState('login');
 
+    const [confirmEmailSent, setConfirmEmailSent] = useState(false);
+
     useEffect( () => {
         db.collection('items').get().then(snapshot => {
             const todos = firebaseLooper(snapshot);
@@ -35,7 +37,7 @@ const App = () => {
     //====Firebase tools====================================
 
     firebase.auth().onAuthStateChanged( userObj => {
-        if (userObj && signInType === 'login') {
+        if (userObj && signInType === 'login' && firebase.auth().currentUser.emailVerified) {
             user = firebase.auth().currentUser;
             setLoggedIn(true);
             //console.log(user);
@@ -51,7 +53,7 @@ const App = () => {
         firebase.auth().signInWithEmailAndPassword('guest@test.com', 'test123')
         .then( () => {
             setLoggedIn(true);
-            console.log(user);
+            //console.log(user);
         })
         .catch(err => console.error(err.message));
     }
@@ -144,7 +146,7 @@ const App = () => {
 
     const contentProps = {
         loggedIn, pendingCount, doneCount, pattern, searchItems, filter, onSwitchFilter, dark, visibleItems, 
-        deleteItem, toggleDone, toggleImportant, addItem, signInType, setSignInType, guestUserSignIn
+        deleteItem, toggleDone, toggleImportant, addItem, signInType, setSignInType, guestUserSignIn, confirmEmailSent, setConfirmEmailSent
     }
 
     return(
